# nodejs-server-server-generated/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Whats App API"
  version: "1.0.0"
  title: "Whats App API"
  contact:
    name: "Irene-Maria Tabakis"
    email: "irma.tabakis@gmail.com"
host: "virtserver.swaggerhub.com"
basePath: "/asterakia/WhatsApp/1.0.0"
...
schemes:
- "https"
paths:
  /users:
    post:
      tags:
      - "user"
      summary: "Create a user object using a phone number."
      description: " "
      operationId: "usersPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Create a user object. Must have the parameter of phone umber\
          \ in order to send the verifcation code."
        required: false
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "A user's phone number is added to the user object. The other\
            \ fields will be null. The other fields will receive a value after the\
            \ user verifies the phone number. The variable name of the response is\
            \ user."
          schema:
            $ref: "#/definitions/User"
        500:
          description: "The phone number was not provided."
      x-swagger-router-controller: "User"
  /users/{userId}:
    get:
      tags:
      - "user"
      summary: "Get info on one specific user."
      operationId: "usersUserIdGET"
      parameters:
      - name: "currentUserId"
        in: "query"
        description: "When querying for a specific user to display user info."
        required: false
        type: "string"
      - name: "userId"
        in: "path"
        description: "The userId of a specific user."
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Returns one user. The variable name of the response is user."
          schema:
            $ref: "#/definitions/inline_response_200"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
    put:
      tags:
      - "user"
      summary: "Update a user"
      description: "Add the mediaId in the body for the user's avatar to be updated"
      operationId: "usersUserIdPUT"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "The id of the user is used to target the specific user the developer\
          \ wishes to edit. The variable name of the response is users."
        required: true
        type: "string"
        minimum: 1
      - in: "body"
        name: "body"
        description: "Edit the user with values found in the request body."
        required: false
        schema:
          $ref: "#/definitions/body"
      responses:
        200:
          description: "A user object will be returned."
          schema:
            $ref: "#/definitions/User"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
    delete:
      tags:
      - "user"
      summary: "Delete a user."
      operationId: "usersUserIdDELETE"
      parameters:
      - name: "userId"
        in: "path"
        description: "Delete a user. Place the ID of the user you wish to delete in\
          \ the path."
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "A user deleted."
          schema:
            type: "string"
            example: "User 'ID number' was deleted"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/friends:
    get:
      tags:
      - "user"
      summary: "Get all the friend's belonging to this user."
      operationId: "usersUserIdFriendsGET"
      parameters:
      - name: "userId"
        in: "path"
        description: "Get all the friend's belonging to this user."
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Returns all user's who are friend with the specified userId\
            \ given. The variable name of the response is users."
          schema:
            $ref: "#/definitions/User"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/blocked:
    get:
      tags:
      - "user"
      summary: "Get all the blocked users belonging to a specifc user."
      operationId: "usersUserIdBlockedGET"
      parameters:
      - name: "userId"
        in: "path"
        description: "The id of the user is needed to query the blocked user's of\
          \ this specifc individual."
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Returns a list of all the blocked user's relating to this\
            \ user. The variable name of the response is users."
          schema:
            $ref: "#/definitions/User"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/messages:
    get:
      tags:
      - "user"
      summary: "Get all messages belonging to a specific user."
      operationId: "usersUserIdMessagesGET"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "All the messages this user has sent will be returned. The\
            \ variable name of the response is messages."
          schema:
            $ref: "#/definitions/Message"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/messages/unread:
    get:
      tags:
      - "user"
      summary: "Get all unread messages belonging to a user."
      operationId: "usersUserIdMessagesUnreadGET"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Each key in the JSON object returned is the chat number assoicated\
            \ with the unread messages located in the array. The variable name of\
            \ the response is messages."
          schema:
            $ref: "#/definitions/inline_response_200_1"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
...
  /users/{userId}/messages/seen:
    get:
      tags:
      - "user"
      summary: "Get all messages belonging to a user that have been seen by another\
        \ user/ other users."
      operationId: "usersUserIdMessagesSeenGET"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Each key in the JSON object returned is the chat number assoicated\
            \ with the messages that have been seen located in the array. The variable\
            \ name of the response is messages."
          schema:
            $ref: "#/definitions/inline_response_200_3"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/messages/favorite:
    get:
      tags:
      - "user"
      summary: "Get all favorite messages belonging to a user."
      operationId: "usersUserIdMessagesFavoriteGET"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "All favorited messages returned. The variable name of the\
            \ response is favoriteMessages."
          schema:
            $ref: "#/definitions/inline_response_200_4"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/calls:
    get:
      tags:
      - "user"
      summary: "Get all calls made by a specific user."
      operationId: "usersUserIdCallsGET"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "All the calls this user has made will be returned. The variable\
            \ name of the response is calls"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/calls/nonanswered:
    get:
      tags:
      - "user"
      summary: "Get all non answered calls belonging to a user."
      operationId: "usersUserIdCallsNonansweredGET"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Each key in the JSON object returned is the chat number assoicated\
            \ with the non answered calls located in the array. The variable name\
            \ of the response is calls."
          schema:
            $ref: "#/definitions/inline_response_200_5"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/calls/answered:
    get:
      tags:
      - "user"
      summary: "Get all answered calls."
      operationId: "usersUserIdCallsAnsweredGET"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Each key in the JSON object returned is the chat number assoicated\
            \ with the call located in the array. The variable name of the response\
            \ is calls."
          schema:
            $ref: "#/definitions/inline_response_200_6"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/private&group_chats:
    get:
      tags:
      - "user"
      summary: "Get both private and group chats and calls the user is a participant\
        \ of"
      operationId: "usersUserIdPrivategroup_chatsGET"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "All the private and group chats will be returned with the\
            \ chat type equivlant to \"private\" or \"group\". There will also be\
            \ a type to every object, the type can be equivlant to \"unreadAndFavorite,\"\
            \ \"unread\", \"favorite\",\"readandNonfavorite\", \"delivered\", \"deliveredAndseen\"\
            , \"answered\" or \"nonanswered\"."
          schema:
            $ref: "#/definitions/inline_response_200_7"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/leave/{chatId}:
    put:
      tags:
      - "user"
      summary: "This route is envoked when a user leaves a chat chat"
      operationId: "usersUserIdLeaveChatIdPUT"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      - name: "chatId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Returns the userChats object. The response variable name will\
            \ be userChats"
          schema:
            $ref: "#/definitions/UserChats"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/check_contacts_use_whatsApp:
    post:
      tags:
      - "user"
      summary: "Check if contacts are WhatApp's users"
      operationId: "usersUserIdCheck_contacts_use_whatsAppPOST"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      - name: "returnUserInfo"
        in: "query"
        description: "if returnUserInfo is set to true, all the info about the user\
          \ will be sent back."
        required: false
        type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "array"
          items:
            $ref: "#/definitions/body_1"
      responses:
        200:
          description: "Returns the users which use WhatApp and the invalid numbers\
            \ the client sent in. If returnUserInfo is set to true, then the entire\
            \ user object will be sent in"
          schema:
            $ref: "#/definitions/inline_response_200_8"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/invite_to_WhatsApp:
    post:
      tags:
      - "user"
      summary: "Invite a contact to download WhatsApp via SMS"
      operationId: "usersUserIdInvite_to_WhatsAppPOST"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "array"
          items:
            $ref: "#/definitions/body_2"
      responses:
        200:
          description: "Sends a message to the numbers provided."
          schema:
            $ref: "#/definitions/inline_response_200_9"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
  /users/{userId}/notifications/global:
    patch:
      tags:
      - "user"
      summary: "Edit the global settings for notifications"
      operationId: "usersUserIdNotificationsGlobalPATCH"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      - name: "unread"
        in: "query"
        description: "This query parameter can either be set to true or false"
        required: false
        type: "string"
      - name: "delivered"
        in: "query"
        description: "This query parameter can either be set to true or false"
        required: false
        type: "string"
      - name: "seen"
        in: "query"
        description: "This query parameter can either be set to true or false"
        required: false
        type: "string"
      - name: "nonanswered"
        in: "query"
        description: "This query parameter can either be set to true or false"
        required: false
        type: "string"
      responses:
        200:
          description: "On sucess, the updated user will be returned. The name of\
            \ the return object is user"
          schema:
            $ref: "#/definitions/User"
      security:
      - Authorization: []
      x-swagger-router-controller: "User"
...
securityDefinitions:
  Authorization:
    type: "apiKey"
    name: "Authorization"
    in: "header"
  Socket:
    type: "apiKey"
    name: "socket"
    in: "header"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "number"
        description: "Unique identifer"
      first_name:
        type: "string"
      last_name:
        type: "string"
      phone:
        type: "string"
      country_code:
        type: "string"
      email:
        type: "string"
      profile_picture_string:
        type: "string"
      avatar:
        type: "string"
      verified:
        type: "boolean"
      passcode:
        type: "string"
      passcode_expiration:
        type: "string"
      socket_id:
        type: "string"
    example:
      country_code: "country_code"
      socket_id: "socket_id"
      passcode_expiration: "passcode_expiration"
      phone: "phone"
      verified: true
      last_name: "last_name"
      id: 0.80082819046101150206595775671303272247314453125
      avatar: "avatar"
      first_name: "first_name"
      email: "email"
      profile_picture_string: "profile_picture_string"
      passcode: "passcode"
...
  UserChats:
    type: "object"
    properties:
      userId:
        type: "string"
      chatId:
        type: "string"
      favorite:
        type: "boolean"
      last_time_in_chat:
        type: "string"
      unread_messages:
        type: "boolean"
      entered_chat_quantity:
        type: "number"
      current_chat:
        type: "boolean"
      message_notifications:
        type: "boolean"
      user_added_notifications:
        type: "boolean"
      messages_not_to_display:
        type: "array"
        example:
        - 50
        - 100
        - 150
        items:
          type: "number"
          description: "The message ID's which will not be displayed to a user because\
            \ they were self-destructing messages"
    example:
      last_time_in_chat: "last_time_in_chat"
      chatId: "chatId"
      entered_chat_quantity: 0.80082819046101150206595775671303272247314453125
      current_chat: true
      messages_not_to_display:
      - 50
      - 100
      - 150
      unread_messages: true
      user_added_notifications: true
      userId: "userId"
      favorite: true
      message_notifications: true
  Message:
    type: "object"
    properties:
      id:
        type: "number"
      message:
        type: "string"
      chatId:
        type: "string"
      userId:
        type: "string"
      type:
        type: "string"
        example: "media, string"
      createdAt:
        type: "string"
        description: "The date time of when the message was created (sent)"
      name:
        type: "string"
        description: "The full name of the user who sent the message"
      post:
        type: "boolean"
      user:
        $ref: "#/definitions/inline_response_200_7_message_user"
      GroupChat:
        $ref: "#/definitions/Message_GroupChat"
      mediaArray:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_7_message_mediaArray"
      files:
        type: "array"
        items:
          $ref: "#/definitions/Message_files"
    example:
      createdAt: "createdAt"
      GroupChat:
        socket_id: "socket_id"
        verified: true
        last_name: "last_name"
        avatar: "avatar"
        profile_picture_string: "profile_picture_string"
        country_code: "country_code"
        passcode_expiration: "passcode_expiration"
        phone: "phone"
        id: 6.02745618307040320615897144307382404804229736328125
        position: "{}"
        first_name: "first_name"
        email: "email"
        passcode: "passcode"
      chatId: "chatId"
      post: true
      name: "name"
      files:
      - s3_url: "s3_url"
        file_string: "file_string"
      - s3_url: "s3_url"
        file_string: "file_string"
      id: 0.80082819046101150206595775671303272247314453125
      message: "message"
      type: "media, string"
      userId: "userId"
      user:
        avatar: "avatar"
        profile_picture_string: "profile_picture_string"
      mediaArray:
      - media_file_string: "media_file_string"
        media_file: "media_file"
      - media_file_string: "media_file_string"
        media_file: "media_file"
...
  inline_response_200:
    properties:
      number_of_favorited_messages:
        type: "number"
        description: "If currentUserId is null, the number of favorited messages will\
          \ be the total number of favorited messages of the queried user. If currentUserId\
          \ is not null, this value will be the number of favorited messages the currentUserId\
          \ favorited of the user being queried."
      user:
        $ref: "#/definitions/inline_response_200_user"
    example:
      number_of_favorited_messages: 0.80082819046101150206595775671303272247314453125
      user:
        socket_id: "socket_id"
        verified: true
        last_name: "last_name"
        avatar: "avatar"
        profile_picture_string: "profile_picture_string"
        country_code: "country_code"
        passcode_expiration: "passcode_expiration"
        relationship_status: "no_relationship, blocked, friends"
        phone: "phone"
        id: 6.02745618307040320615897144307382404804229736328125
        first_name: "first_name"
        email: "email"
        passcode: "passcode"
  body:
    type: "object"
    properties:
      mediaId:
        type: "string"
      id:
        type: "number"
        description: "Unique identifer"
      first_name:
        type: "string"
      last_name:
        type: "string"
      phone:
        type: "string"
      country_code:
        type: "string"
      email:
        type: "string"
      profile_picture_string:
        type: "string"
      avatar:
        type: "string"
      verified:
        type: "boolean"
      passcode:
        type: "string"
      passcode_expiration:
        type: "string"
      socket_id:
        type: "string"
...
  inline_response_200_1:
    properties:
      unreadMessages:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_1_unreadMessages"
    example:
      unreadMessages:
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        userAvatar: "userAvatar"
        id: 0.80082819046101150206595775671303272247314453125
        message: "message"
        type: "type"
        userId: 6.02745618307040320615897144307382404804229736328125
        mediaArray:
        - "mediaArray"
        - "mediaArray"
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        userAvatar: "userAvatar"
        id: 0.80082819046101150206595775671303272247314453125
        message: "message"
        type: "type"
        userId: 6.02745618307040320615897144307382404804229736328125
        mediaArray:
        - "mediaArray"
        - "mediaArray"
...
  inline_response_200_3:
    properties:
      seenMessages:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_1_unreadMessages"
    example:
      seenMessages:
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        userAvatar: "userAvatar"
        id: 0.80082819046101150206595775671303272247314453125
        message: "message"
        type: "type"
        userId: 6.02745618307040320615897144307382404804229736328125
        mediaArray:
        - "mediaArray"
        - "mediaArray"
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        userAvatar: "userAvatar"
        id: 0.80082819046101150206595775671303272247314453125
        message: "message"
        type: "type"
        userId: 6.02745618307040320615897144307382404804229736328125
        mediaArray:
        - "mediaArray"
        - "mediaArray"
  inline_response_200_4:
    properties:
      favoriteMessages:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_1_unreadMessages"
    example:
      favoriteMessages:
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        userAvatar: "userAvatar"
        id: 0.80082819046101150206595775671303272247314453125
        message: "message"
        type: "type"
        userId: 6.02745618307040320615897144307382404804229736328125
        mediaArray:
        - "mediaArray"
        - "mediaArray"
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        userAvatar: "userAvatar"
        id: 0.80082819046101150206595775671303272247314453125
        message: "message"
        type: "type"
        userId: 6.02745618307040320615897144307382404804229736328125
        mediaArray:
        - "mediaArray"
        - "mediaArray"
...
  inline_response_200_5:
    properties:
      nonAnsweredCalls:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_5_nonAnsweredCalls"
    example:
      nonAnsweredCalls:
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        userAvatar: "userAvatar"
        id: 0.80082819046101150206595775671303272247314453125
        userId: 6.02745618307040320615897144307382404804229736328125
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        userAvatar: "userAvatar"
        id: 0.80082819046101150206595775671303272247314453125
        userId: 6.02745618307040320615897144307382404804229736328125
  inline_response_200_6:
    properties:
      answeredCalls:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_5_nonAnsweredCalls"
    example:
      answeredCalls:
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        userAvatar: "userAvatar"
        id: 0.80082819046101150206595775671303272247314453125
        userId: 6.02745618307040320615897144307382404804229736328125
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        userAvatar: "userAvatar"
        id: 0.80082819046101150206595775671303272247314453125
        userId: 6.02745618307040320615897144307382404804229736328125
...
  inline_response_200_7:
    properties:
      success:
        type: "boolean"
      chats:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_7_chats"
    example:
      success: true
      chats:
      - createdAt: "createdAt"
        last_activity: "last_activity"
        groupChatInfo:
          chatId: 5.962133916683182377482808078639209270477294921875
          name: "name"
          adminId: 1.46581298050294517310021547018550336360931396484375
          avatar: "avatar"
          id: 6.02745618307040320615897144307382404804229736328125
          profile_picture_string: "profile_picture_string"
        participantIds:
        - 0.80082819046101150206595775671303272247314453125
        - 0.80082819046101150206595775671303272247314453125
        type: "unreadAndFavorite, unread, favorite, readandNonfavorite, delivered,\
          \          deliveredAndseen, answered, nonanswered"
        message:
          call: "call"
          createdAt: "createdAt"
          chatId: "chatId"
          name: "name"
          files:
          - file_string: "file_string"
            url: "url"
          - file_string: "file_string"
            url: "url"
          id: 5.63737665663332876420099637471139430999755859375
          message: "message"
          type: "media, string"
          userId: "userId"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
          mediaArray:
          - media_file_string: "media_file_string"
            media_file: "media_file"
          - media_file_string: "media_file_string"
            media_file: "media_file"
        user:
          last_name: "last_name"
          avatar: "avatar"
          first_name: "first_name"
          profile_picture_string: "profile_picture_string"
        chatType: "group or private"
      - createdAt: "createdAt"
        last_activity: "last_activity"
        groupChatInfo:
          chatId: 5.962133916683182377482808078639209270477294921875
          name: "name"
          adminId: 1.46581298050294517310021547018550336360931396484375
          avatar: "avatar"
          id: 6.02745618307040320615897144307382404804229736328125
          profile_picture_string: "profile_picture_string"
        participantIds:
        - 0.80082819046101150206595775671303272247314453125
        - 0.80082819046101150206595775671303272247314453125
        type: "unreadAndFavorite, unread, favorite, readandNonfavorite, delivered,\
          \          deliveredAndseen, answered, nonanswered"
        message:
          call: "call"
          createdAt: "createdAt"
          chatId: "chatId"
          name: "name"
          files:
          - file_string: "file_string"
            url: "url"
          - file_string: "file_string"
            url: "url"
          id: 5.63737665663332876420099637471139430999755859375
          message: "message"
          type: "media, string"
          userId: "userId"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
          mediaArray:
          - media_file_string: "media_file_string"
            media_file: "media_file"
          - media_file_string: "media_file_string"
            media_file: "media_file"
        user:
          last_name: "last_name"
          avatar: "avatar"
          first_name: "first_name"
          profile_picture_string: "profile_picture_string"
        chatType: "group or private"
  body_1:
    properties:
      phone:
        type: "string"
        description: "Only include the country code and digits of the number in the\
          \ phone string"
...
  body_2:
    properties:
      phone:
        type: "string"
        example: "3105910213"
...